# money-miner-api/openapi.yaml
openapi: 3.0.3
info:
  title: money-miner API
  version: 0.1.0
  description: |
    REST + realtime contract for the money-miner master server.

    Auth: Keycloak RS256 bearer tokens (`Authorization: Bearer <jwt>`),
    roles from `realm_access.roles`. Every user-owned resource is scoped by
    the JWT `sub` (tenant isolation). Errors use the envelope
    `{"error": {"code", "message", "details?}}`.

    Realtime: SSE `GET /v1/events` for user-facing live updates; WebSocket
    `/v1/swarm/ws` is reserved for the worker control channel.

    Honesty: objects produced by the SIMULATED engine always carry
    `"engine": "simulated"`. There are no balances and no withdrawals —
    payouts are on-chain, explorer-linked observations.
  contact:
    name: money-miner
    url: https://github.com/thecsdoctor/money-miner
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.txt
servers:
  - url: https://money-miner.thecsdoctor.com/api
    description: production (via nginx edge; /api rewrites to /v1)
  - url: http://127.0.0.1:8080/v1
    description: local loopback
...
paths:
...
  /currencies:
    get:
      tags: [currencies]
      summary: Coin catalog (20 rows)
      operationId: listCurrencies
      responses:
        '200':
          description: catalog
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Currency' }
                  total: { type: integer, example: 20 }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /currencies/{symbol}:
    get:
      tags: [currencies]
      summary: Currency detail incl. adapter_config, pools, explorer templates
      operationId: getCurrency
      parameters:
        - { name: symbol, in: path, required: true, schema: { type: string, example: XMR } }
      responses:
        '200':
          description: currency
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Currency' }
        '404': { $ref: '#/components/responses/NotFound' }
  /wallets:
    get:
      tags: [wallets]
      summary: My wallets
      operationId: listWallets
      responses:
        '200':
          description: wallets
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Wallet' } }
                  total: { type: integer }
    post:
      tags: [wallets]
      summary: Attach a payout wallet (format-validated; addresses only, never keys)
      operationId: createWallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currency, address]
              properties:
                currency: { type: string, example: XMR }
                address: { type: string }
                label: { type: string }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Wallet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { description: wallet for this currency already exists, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /wallets/{id}:
    delete:
      tags: [wallets]
      summary: Delete a wallet
      operationId: deleteWallet
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '204': { description: deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /miners:
    get:
      tags: [miners]
      summary: List my miners (with live hashrate)
      operationId: listMiners
      responses:
        '200':
          description: miners
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Miner' } }
                  total: { type: integer }
    post:
      tags: [miners]
      summary: Create a miner
      description: Rejects `engine_strategy=deferred` currencies (XCH) with `currency_deferred`.
      operationId: createMiner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, currency, pool_url]
              properties:
                name: { type: string }
                currency: { type: string, example: XMR }
                wallet_id: { type: string, format: uuid }
                pool_url: { type: string, example: 'stratum+tcp://pool.supportxmr.com:443' }
                engine: { type: string, enum: [adapter, native-go, simulated], description: 'simulated only when the server has ALLOW_SIMULATED=true' }
                cpu_pct: { type: integer, minimum: 1, maximum: 100, default: 50 }
                gpu_pct: { type: integer, minimum: 0, maximum: 100, default: 0 }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Miner' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /miners/{id}:
    get:
      tags: [miners]
      summary: Miner detail + current run
      operationId: getMiner
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: miner
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MinerDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [miners]
      summary: Rename / re-wallet / re-pool a miner
      operationId: updateMiner
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                wallet_id: { type: string, format: uuid }
                pool_url: { type: string }
      responses:
        '200':
          description: updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Miner' }
    delete:
      tags: [miners]
      summary: Delete a miner (runs/payouts history survives)
      operationId: deleteMiner
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '204': { description: deleted }
...
  /miners/{id}/allocation:
    post:
      tags: [miners]
      summary: Live allocation update
      description: 'Clamps: native workers honor threads=int(cores*cpu_pct); browser workers clamp to <=50% duty cycle regardless.'
      operationId: setMinerAllocation
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cpu_pct: { type: integer, minimum: 1, maximum: 100 }
                gpu_pct: { type: integer, minimum: 0, maximum: 100 }
      responses:
        '200':
          description: updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Miner' }
...
  /workers/{id}:
    get:
      tags: [swarm]
      summary: Worker detail + recent metrics
      operationId: getWorker
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: worker
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Worker' }
    delete:
      tags: [swarm]
      summary: Delete a worker record
      operationId: deleteWorker
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '204': { description: deleted }
  /workers/{id}/revoke:
    post:
      tags: [swarm]
      summary: Revoke a worker token (live WS closed, engines stopped on next contact)
      operationId: revokeWorker
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: revoked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Worker' }
...
  /payouts:
    get:
      tags: [payouts]
      summary: Payout ledger with explorer links
      operationId: listPayouts
      parameters:
        - { name: currency, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: payouts
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Payout' } }
                  total: { type: integer }
...
  /settings/{key}:
    put:
      tags: [settings]
      summary: Set a setting (theme, language, electricity_cost, ...)
      operationId: putSetting
      parameters:
        - { name: key, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: { }
      responses:
        '200': { description: stored }
...
  /currency-settings/{symbol}:
    put:
      tags: [settings]
      summary: Set per-currency override (pool_url, enabled, custom_adapter)
      operationId: putCurrencySetting
      parameters:
        - { name: symbol, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pool_url: { type: string }
                enabled: { type: boolean }
                custom_adapter: { type: object, additionalProperties: true }
      responses:
        '200': { description: stored }
...
  /exchange-settings/{exchange}:
    put:
      tags: [settings]
      summary: Set exchange preference (optional read-only API key, pgcrypto-encrypted at rest)
      operationId: putExchangeSetting
      parameters:
        - { name: exchange, in: path, required: true, schema: { type: string, example: kraken } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currencies: { type: array, items: { type: string } }
                api_key: { type: string, description: 'read-only key; encrypted at rest; never returned by the API' }
      responses:
        '200': { description: stored }
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak RS256 access token; tenant = `sub`; roles in `realm_access.roles`.
  responses:
    Unauthorized:
      description: missing/invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: validation failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Currency:
      type: object
      required: [symbol, name, algorithm, hardware_class, engine_strategy, enabled]
      properties:
        id: { type: string, format: uuid }
        symbol: { type: string, example: XMR }
        name: { type: string, example: Monero }
        algorithm: { type: string, example: randomx }
        hardware_class: { type: string, enum: [cpu, gpu, cpu+gpu, storage] }
        engine_strategy: { type: string, enum: [adapter, native-go, browser-wasm, deferred] }
        adapter_config: { type: object, additionalProperties: true }
        min_vram_mb: { type: integer, nullable: true }
        explorer_tx_tpl: { type: string, example: 'https://xmrchain.net/tx/{txid}' }
        explorer_addr_tpl: { type: string }
        homepage: { type: string }
        pools:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              url: { type: string }
              api_tpl: { type: string, description: 'pool stats API URL template, {address} placeholder' }
              fee_pct: { type: number }
        enabled: { type: boolean }
        deferred_note: { type: string, description: 'present when engine_strategy=deferred' }
...
security:
  - bearerAuth: []